+++ GameSelector.h
@@ -9,8 +9,6 @@
 
 #include "Program.h"
 
-// #define DEBUGGING_MENU 1
-
 // Just as a reminder coord len shouldnt be less than 2
 #define COORDINATES_LENGTH 12
 #define NAN 64
@@ -38,10 +36,10 @@ public:
 
 	virtual void onClick();
 	virtual void onDoubleClick();
-	virtual void onLeftGesture(uint8_t);
-	virtual void onRightGesture(uint8_t);
-	virtual void onUpGesture(uint8_t);
-	virtual void onDownGesture(uint8_t);
+	virtual void onLeftGesture(unsigned int);
+	virtual void onRightGesture(unsigned int);
+	virtual void onUpGesture(unsigned int);
+	virtual void onDownGesture(unsigned int);
 	virtual void onButtonPressed();
 
 	virtual uint8_t run();
